ConfirmValue toegevoegd aan priceMenu

+++ Project/Presentation/PriceMenu.cs
@@ -66,7 +66,7 @@ public static class PriceMenu
 
     }
 
-    static void DisplayOptions(int selectedOption)
+    public static void DisplayOptions(int selectedOption)
     {
         Console.WriteLine("Selecteer een optie:");
 
@@ -78,7 +78,7 @@ public static class PriceMenu
         // Display option 2
         Console.ForegroundColor = selectedOption == 2 ? ConsoleColor.Green : ConsoleColor.White;
         Console.Write(selectedOption == 2 ? ">> " : "   ");
-        Console.WriteLine("[2] Een prijscategorie editen.");
+        Console.WriteLine("[2] Een prijscategorie updaten.");
 
         // Display option 3
         Console.ForegroundColor = selectedOption == 3 ? ConsoleColor.Green : ConsoleColor.White;
@@ -108,8 +108,9 @@ public static class PriceMenu
 
         if (priceModel != null)
         {
+            Console.Clear();
             ShowPriceInformation(priceModel);
-            Console.WriteLine("Wilt u deze prijscategorie verwijderen? Y/N: ");
+            Console.WriteLine("Wilt u deze prijscategorie verwijderen? J/N: ");
             string? answer = Console.ReadLine();
             // if (Helper.IsString(answer))
             // {
@@ -120,20 +121,29 @@ public static class PriceMenu
                 Console.WriteLine("Uw antwoord is nee.");
                 BackToStartMenu();
             }
-            else if (answer!= null && answer.ToLower() != "y")
+            else if (answer!= null && answer.ToLower() != "j")
             {
                 Console.WriteLine("Verkeerde input!");
                 Thread.Sleep(3000);
                 DeletePriceCategory();
             }
-
-            pricesLogic.DeletePriceCategory(priceModel.Id);
-            Console.WriteLine("De verwijdering is voltooid");
+            
+            if (ConfirmValue(priceModel, null, false, true))
+            {
+                pricesLogic.DeletePriceCategory(priceModel.Id);
+            }
+            else
+            {
+                BackToStartMenu();
+            }
         }
         else
         {
+            Console.ForegroundColor = ConsoleColor.Red;
             Console.WriteLine("Geen prijscategorie gevonden");
             Thread.Sleep(3000);
+            Console.ResetColor();
+
         }
         BackToStartMenu();
 
@@ -142,29 +152,66 @@ public static class PriceMenu
     {
         Console.WriteLine("Hieronder kunt u de huidige prijscategorieën zien");
         ShowAllPricesInformation();
-        Console.WriteLine("Wilt u nog een prijscategorie toevoegen? J/N: ");
-        string? answer = Console.ReadLine();
-
-        if (answer!= null && answer.ToLower() == "n")
+        while (true)
         {
-            Console.WriteLine("Uw antwoord is nee.");
-            BackToStartMenu();
-        }
-        else if (answer!= null && answer.ToLower() != "j")
-        {
-            Console.WriteLine("Verkeerde input!");
-            Thread.Sleep(3000);
-            AddPriceCategory();
-        }
+            Console.WriteLine("Wilt u nog een prijscategorie toevoegen? J/N: ");
+            string? answer = Console.ReadLine();
+            if (answer!= null && answer.ToLower() == "n")
+            {
+                Console.WriteLine("Uw antwoord is nee.");
+                BackToStartMenu();
+                break;
+            }
+            else if (answer!= null && answer.ToLower() == "j" )
+            {
+                Thread.Sleep(1000);
+                break;
 
+            }
+            else if (answer!= null && answer.ToLower() != "j")
+            {
+                Console.WriteLine("Verkeerde input!");
+                Thread.Sleep(2000);
+            }
+        }
+        
         Console.WriteLine("Voer de naam van het nieuwe passenger type in: ");
-        string? NewPassenger = Console.ReadLine();
+        string? newPassenger = Console.ReadLine();
         Console.WriteLine("Bepaal de prijs voor de nieuwe passenger: ");
-        double NewPrice = Convert.ToDouble(Console.ReadLine());
+        string newPrice = Console.ReadLine();
+        PriceModel newPriceCategory = new (pricesLogic.GenerateNewId(), newPassenger, Convert.ToDouble(newPrice));
+        if (ConfirmValue(newPriceCategory, newPrice, false))
+        {
+            pricesLogic.UpdateList(newPriceCategory);
+            Console.WriteLine($"De prijscategorie '{newPassenger}' is toevoegd");
+        }
+        else
+        {
+            while (true)
+            {
+                Console.WriteLine("Wilt u nog een prijscategorie toevoegen? J/N: ");
+                string? answer = Console.ReadLine();
+                if (answer!= null && answer.ToLower() == "n")
+                {
+                    Console.WriteLine("Uw antwoord is nee.");
+                    BackToStartMenu();
+                    break;
+                }
+                else if (answer!= null && answer.ToLower() == "j" )
+                {
+                    Thread.Sleep(1000);
+                    AddPriceCategory();
+                    break;
 
-        PriceModel NewPriceCategory = new (pricesLogic.GenerateNewId(), NewPassenger, NewPrice);
-        pricesLogic.UpdateList(NewPriceCategory);
-        Console.WriteLine($"De prijscategorie '{NewPassenger}' is toevoegd");
+                }
+                else if (answer!= null && answer.ToLower() != "j")
+                {
+                    Console.WriteLine("Verkeerde input!");
+                    Thread.Sleep(2000);
+                }
+            }
+        }
+
         Thread.Sleep(2000);
         BackToStartMenu();
     }
@@ -177,14 +224,14 @@ public static class PriceMenu
         if (priceModel != null)
         {
             ShowPriceInformation(priceModel);
-            Console.WriteLine("Wilt u deze prijscategorie bewerken? Y/N: ");
+            Console.WriteLine("Wilt u deze prijscategorie bewerken? J/N: ");
             string? answer = Console.ReadLine();
             if (answer!= null && answer.ToLower() == "n")
             {
                 Console.WriteLine("Uw antwoord is nee.");
                 BackToStartMenu();
             }
-            else if (answer!= null && answer.ToLower() != "y")
+            else if (answer!= null && answer.ToLower() != "j")
             {
                 Console.WriteLine("Verkeerde input!");
                 Thread.Sleep(3000);
@@ -194,7 +241,12 @@ public static class PriceMenu
         }
         else
         {
-            Console.WriteLine("Geen prijscategorie gevonden");
+            Console.ForegroundColor = ConsoleColor.Red;
+            Console.WriteLine("Geen prijscategorie gevonden!");
+            Thread.Sleep(3000);
+            Console.Clear();
+            Console.ResetColor();
+            EditPriceCategory();
         }
         BackToStartMenu();
     
@@ -218,8 +270,36 @@ public static class PriceMenu
         // priceModel.Passenger = NewPassenger;
 
         Console.WriteLine("Bepaal de nieuwe prijs: ");
-        double NewPrice = Convert.ToDouble(Console.ReadLine());
-        priceModel.Price = NewPrice;
+        string NewPrice = Console.ReadLine();
+        if (ConfirmValue(null, NewPrice, true))
+        {
+            priceModel.Price = Convert.ToDouble(NewPrice);
+        }
+        else
+        {
+            while (true)
+            {
+                Console.WriteLine("Wilt u verder met het toevoegen J/N");
+                string answer = Console.ReadLine();
+                if (answer!= null && answer.ToLower() == "n")
+                {
+                    Console.WriteLine("Uw antwoord is nee.");
+                    BackToStartMenu();
+                    break;
+                }
+                else if (answer!= null && answer.ToLower() == "j")
+                {
+                    Thread.Sleep(1000);
+                    UpdatePriceCategory(priceModel);
+                    break;
+                }
+                else if (answer!= null && answer.ToLower() != "j")
+                {
+                    Console.WriteLine("Verkeerde input!");
+                    Thread.Sleep(2000);
+                }
+            }
+        }
 
         pricesLogic.UpdateList(priceModel);
         Console.WriteLine("De bewerking is voltooid, hier is het resultaat: ");
@@ -288,5 +368,75 @@ public static class PriceMenu
         var passenger = priceModel.Passenger;
         var price = priceModel.Price;
         return new List<string> { $"{id}", $"{passenger}", $"{price}" };
+    }
+
+    public static bool ConfirmValue(PriceModel priceModel, string UpdatedValue = null, bool IsUpdate = false, bool delete = false)
+    {
+        if (IsUpdate && string.IsNullOrEmpty(UpdatedValue) && !delete || !IsUpdate && (priceModel == null) && !delete)
+        {
+            Console.ForegroundColor = ConsoleColor.Red;
+            Console.WriteLine(IsUpdate ? "Ongeldige invoer." : "Fout: Nieuwe prijsgevens ontbreken!");
+            Console.ResetColor();
+            Thread.Sleep(2000);
+            Console.Clear();
+            return false;
+        }
+
+        if (delete)
+        {
+            Console.WriteLine($"U staat op het punt de prijscategorie te verwijderen met de volgende info");
+            ShowPriceInformation(priceModel);
+        }
+        else if (!IsUpdate)
+        {
+            Console.WriteLine($"U staat op het punt een nieuwe prijscategorie toe te voegen met de volgende info");
+            ShowPriceInformation(priceModel);
+        }
+        else if (IsUpdate)
+        {
+            Console.WriteLine($"U staat op het punt oude data te veranderen: {UpdatedValue}");
+        }
+
+        do
+        {
+            ConsoleKeyInfo keyInfo;
+            Console.Write("Druk op ");
+            Console.ForegroundColor = ConsoleColor.Green;
+            Console.Write("Enter");
+            Console.ResetColor();
+            Console.Write(" om door te gaan of druk op ");
+            Console.ForegroundColor = ConsoleColor.Red;
+            Console.Write("Backspace");
+            Console.ResetColor();
+            Console.WriteLine(" om te annuleren.");
+
+            keyInfo = Console.ReadKey(true);
+            if (keyInfo.Key == ConsoleKey.Backspace)
+            {
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine(!delete ? "Toevoegen geannuleerd." : "Verwijderen geannuleerd");
+                Console.ResetColor();
+                Thread.Sleep(2000);
+                Console.Clear();
+                return false;
+            }
+            else if (keyInfo.Key == ConsoleKey.Enter)
+            {
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(!delete ? "Data is toegevoegd!" : "De verwijdering is voltooid");
+                Console.ResetColor();
+                Thread.Sleep(2000);
+                Console.Clear();
+                return true;
+            }
+            else
+            {
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("Ongeldige invoer!");
+                Console.ResetColor();
+                Thread.Sleep(2000);
+                // return false;
+            }
+        }while(true);
     }
 }
